feat: split entity and dto, minor fixes
