+++ api/apps/api/src/modules/scenarios/locks/entity/scenario.lock.api.entity.ts
@@ -1,4 +1,3 @@
-import { ApiProperty } from '@nestjs/swagger';
 import { CreateDateColumn, Entity, PrimaryColumn } from 'typeorm';
 
 @Entity(`scenario_locks`)
@@ -22,9 +21,4 @@ export class ScenarioLockEntity {
     default: 'now()',
   })
   createdAt!: Date;
-}
-
-export class ScenarioLockResult {
-  @ApiProperty()
-  data!: ScenarioLockEntity;
 }
